# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-grade Go web server with AWS Cognito authentication and AWS service integration.",
        "title": "AWS Go Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/pmollerus23/go-aws-server",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/aws/dynamodb/tables": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all DynamoDB tables in the AWS account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aws"
                ],
                "summary": "List DynamoDB tables",
                "responses": {
                    "200": {
                        "description": "tables and count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to list DynamoDB tables",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/aws/s3/buckets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all S3 buckets in the AWS account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aws"
                ],
                "summary": "List S3 buckets",
                "responses": {
                    "200": {
                        "description": "buckets and count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to list S3 buckets",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/items": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all items in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "List all items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_pmollerus23_go-aws-server_internal_handlers.Item"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new item with name and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create a new item",
                "parameters": [
                    {
                        "description": "Item to create",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pmollerus23_go-aws-server_internal_handlers.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_pmollerus23_go-aws-server_internal_handlers.CreateItemResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ValidationError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_pmollerus23_go-aws-server_internal_handlers.CreateItemRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Item description"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "New Item"
                }
            }
        },
        "github_com_pmollerus23_go-aws-server_internal_handlers.CreateItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Item description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "New Item"
                }
            }
        },
        "github_com_pmollerus23_go-aws-server_internal_handlers.Item": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is a sample item description"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Sample Item"
                }
            }
        },
...
        "handlers.ValidationError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                },
                "problems": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token authentication via AWS Cognito. Use format: \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}